+++ shoppingmall/src/App.js
@@ -7,11 +7,10 @@ import AppLayout from "./components/commom/AppLayout";
 // pages
 import MyPage from "./pages/myPage/MyPage";
 import ProductListPage from "./pages/productPage/ProductListPage";
-import ProductWritePage from "./pages/productPage/ProductWritePage";
 import Login from "../src/components/userAccount/Login";
 import Signup from "./components/userAccount/Signup";
 import Sell from "./pages/sellPage/SellPage";
-import ProductRegistration from "./pages/ProductRegistrationPage/ProductRegistrationPage";
+import Write from "./pages/ProductWritePage/ProductWritePage";
 
 function App() {
   return (
@@ -20,12 +19,11 @@ function App() {
       <AppLayout>
         <Routes>
           <Route path="/settings" element={<MyPage />} />
-          <Route path="/" element={<ProductListPage/>}/>
-          <Route path="/write"  element={<ProductWritePage/>}/>
+          <Route path="/" element={<ProductListPage />} />
+          <Route path="/write" element={<Write />} />
           <Route path="/login" element={<Login />} />
           <Route path="/signup" element={<Signup />} />
           <Route path="/sell" element={<Sell />} />
-          <Route path="/PR" element={<ProductRegistration />} />
         </Routes>
       </AppLayout>
     </ThemeProvider>
