+++ shoppingmall/src/pages/productPage/ProductListPage.js
@@ -1,277 +1,278 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
+import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import styled from "styled-components";
 import { theme } from "../../style/theme";
-import ModalComponent from '../../components/modal/ModalComponent';
-import Pagination from "../../components/commom/Pagination";
+import ModalComponent from "../../components/modal/ModalComponent";
+import Pagination from "../../components/sellpage/PaginationArea";
 // svg
 import unlike from "../../assets/unlike.svg";
-import pencil from "../../assets/pencil.svg"
+import pencil from "../../assets/pencil.svg";
 
 const ProductListPage = () => {
-   const navigate = useNavigate();
-   const [btnActive, setBtnActive] = useState(1);
-   const [isVisible, setIsVisible] = useState(false);
-   const [isChecked, setIsLike] = useState([]);
-   const [isLogin, setIsLogin] = useState(true);
-   const productsPerRow = 4;
-   const pages = [1, 2, 3, 4, 5];
-   const [currentPage, setCurrentPage] = useState(1);
+  const navigate = useNavigate();
+  const [btnActive, setBtnActive] = useState(1);
+  const [isVisible, setIsVisible] = useState(false);
+  const [isChecked, setIsLike] = useState([]);
+  const [isLogin, setIsLogin] = useState(true);
+  const productsPerRow = 4;
+  const pages = [1, 2, 3, 4, 5];
+  const [currentPage, setCurrentPage] = useState(1);
 
-   const [productList, setProductList] = useState([
+  const [productList, setProductList] = useState([
     {
-        id: 1,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-      {
-        id: 2,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-      {
-        id: 3,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-      {
-        id: 4,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-      {
-        id: 5,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-      {
-        id: 6,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-      {
-        id: 7,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-      {
-        id: 8,
-        image: "https://via.placeholder.com/150",
-        productName: "아디다스",
-        productPrice: "56,000",
-        description:
-          "상품 설명 상품 설명 상품 설명 상품 설명",
-      },
-   ]);
+      id: 1,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+    {
+      id: 2,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+    {
+      id: 3,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+    {
+      id: 4,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+    {
+      id: 5,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+    {
+      id: 6,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+    {
+      id: 7,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+    {
+      id: 8,
+      image: "https://via.placeholder.com/150",
+      productName: "아디다스",
+      productPrice: "56,000",
+      description: "상품 설명 상품 설명 상품 설명 상품 설명",
+    },
+  ]);
 
-   const closeModal = () => {
-        setIsVisible(false);
-   };
+  const closeModal = () => {
+    setIsVisible(false);
+  };
 
-   const onPageChange = (page) => {
-        setCurrentPage(page);
-        console.log(`Page changed to ${page}`);
-    };
+  const onPageChange = (page) => {
+    setCurrentPage(page);
+    console.log(`Page changed to ${page}`);
+  };
 
-   const checkLogin = () => {
-       if (isLogin === true) {
-        navigate("/PR");
-        setIsVisible(false);
-       } else {
-        setIsVisible(true);
-       }
-   };
+  const checkLogin = () => {
+    if (isLogin === true) {
+      navigate("/PR");
+      setIsVisible(false);
+    } else {
+      setIsVisible(true);
+    }
+  };
 
-   const handleClickButton = idx => {
+  const handleClickButton = (idx) => {
     setBtnActive(idx);
-   };
+  };
 
-   const handleHeart = (index) => {
+  const handleHeart = (index) => {
     setIsLike(() => {
       const newLike = [...isChecked];
       newLike[index] = !newLike[index];
-      return newLike;  
-    })
-   }
+      return newLike;
+    });
+  };
 
-    return (
-        <ProductListPageWrapper >
-            <ProductListFilterWrapper>
-                <LeftWrapper>
-                    <ButtonDivs>
-                        <Menu 
-                            onClick={() => handleClickButton(0)}
-                            isActive={btnActive === 0}>
-                            추천순    
-                        </Menu>
-                        <Menu 
-                            onClick={() => handleClickButton(1)}
-                            isActive={btnActive === 1}>
-                            가격 낮은순    
-                        </Menu>
-                        <Menu 
-                            onClick={() => handleClickButton(2)}
-                            isActive={btnActive === 2}>
-                            가격 높은순    
-                        </Menu>
-                    </ButtonDivs>
-                    <Selector name="category">
-                        <option value="cloth">의류</option>
-                        <option value="home">가전</option>
-                        <option value="digital">디지털</option>
-                    </Selector>
-                </LeftWrapper>
-                <RightWrapper>
-                    <button onClick={checkLogin}>상품 등록 &nbsp;
-                            <Icon src={pencil}/>
-                    </button>
-               </RightWrapper>
-            </ProductListFilterWrapper>
-            <ProductListContainer>
-                {productList.map((product, index) => (
-                    <ProductItem key={index} productsPerRow = { productsPerRow} >
-                        <ProductImage src={product.image} alt={product.productName}/>
-                        <ProductInfoWrapper>
-                            <div>
-                                <ProductName>{product.productName}</ProductName>
-                                <ProductPrice>{product.productPrice + " 원"}</ProductPrice>
-                                <ProductDescription>{product.description}</ProductDescription>
-                            </div>
-                            <IconWrapper onClick={() => {handleHeart(index)}}>
-                                    <Icon src={unlike} fill="grayBgColor"/>
-                            </IconWrapper>
-                        </ProductInfoWrapper>
-                    </ProductItem>
-                ))}     
-            </ProductListContainer>
-            <Pagination
-                pages={pages}
-                currentPage={currentPage}
-                onPageChange={onPageChange}
-            />
-           
-            {isVisible && (
-                <ModalComponent 
-                    title="로그인이 필요한 기능입니다." 
-                    subText="로그인 페이지로 이동하시겠습니까?" 
-                    urlPath="/login" 
-                    isClosed={closeModal}/>
-                )}     
-        </ProductListPageWrapper>
-    );
+  return (
+    <ProductListPageWrapper>
+      <ProductListFilterWrapper>
+        <LeftWrapper>
+          <ButtonDivs>
+            <Menu
+              onClick={() => handleClickButton(0)}
+              isActive={btnActive === 0}
+            >
+              추천순
+            </Menu>
+            <Menu
+              onClick={() => handleClickButton(1)}
+              isActive={btnActive === 1}
+            >
+              가격 낮은순
+            </Menu>
+            <Menu
+              onClick={() => handleClickButton(2)}
+              isActive={btnActive === 2}
+            >
+              가격 높은순
+            </Menu>
+          </ButtonDivs>
+          <Selector name="category">
+            <option value="cloth">의류</option>
+            <option value="home">가전</option>
+            <option value="digital">디지털</option>
+          </Selector>
+        </LeftWrapper>
+        <RightWrapper>
+          <button onClick={checkLogin}>
+            상품 등록 &nbsp;
+            <Icon src={pencil} />
+          </button>
+        </RightWrapper>
+      </ProductListFilterWrapper>
+      <ProductListContainer>
+        {productList.map((product, index) => (
+          <ProductItem key={index} productsPerRow={productsPerRow}>
+            <ProductImage src={product.image} alt={product.productName} />
+            <ProductInfoWrapper>
+              <div>
+                <ProductName>{product.productName}</ProductName>
+                <ProductPrice>{product.productPrice + " 원"}</ProductPrice>
+                <ProductDescription>{product.description}</ProductDescription>
+              </div>
+              <IconWrapper
+                onClick={() => {
+                  handleHeart(index);
+                }}
+              >
+                <Icon src={unlike} fill="grayBgColor" />
+              </IconWrapper>
+            </ProductInfoWrapper>
+          </ProductItem>
+        ))}
+      </ProductListContainer>
+      <Pagination
+        pages={pages}
+        currentPage={currentPage}
+        onPageChange={onPageChange}
+      />
+
+      {isVisible && (
+        <ModalComponent
+          title="로그인이 필요한 기능입니다."
+          subText="로그인 페이지로 이동하시겠습니까?"
+          urlPath="/login"
+          isClosed={closeModal}
+        />
+      )}
+    </ProductListPageWrapper>
+  );
 };
 
 export default ProductListPage;
 
 const ProductListPageWrapper = styled.div`
-    width: 100%;
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 20px; 
+  width: 100%;
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 20px;
 `;
 
 const ProductListFilterWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
 `;
 
-const LeftWrapper  = styled.div`
-    margin-left: 1%;
-    display: flex;
-    justify-content: flex;
+const LeftWrapper = styled.div`
+  margin-left: 1%;
+  display: flex;
+  justify-content: flex;
 `;
 
 const ButtonDivs = styled.div`
-    display: flex;
-    width: 100%;
-    justify-content: space-around;
+  display: flex;
+  width: 100%;
+  justify-content: space-around;
 `;
 
 const Menu = styled.button`
-    background-color:  ${props => (props.isActive ? "#EB4646" : "")};
-    color:  ${props => (props.isActive ? "#FFFFFF" : "")};
-    width: 80px;
-    height: 30px;
-    border: 1px solid ${theme.border};
-    border-radius: 15px;
-    font-size: 12px;
-    cursor: pointer;
-    &:hover {
-        background-color: ${theme.border};
-        color: ${theme.black};
-      }
+  background-color: ${(props) => (props.isActive ? "#EB4646" : "")};
+  color: ${(props) => (props.isActive ? "#FFFFFF" : "")};
+  width: 80px;
+  height: 30px;
+  border: 1px solid ${theme.border};
+  border-radius: 15px;
+  font-size: 12px;
+  cursor: pointer;
+  &:hover {
+    background-color: ${theme.border};
+    color: ${theme.black};
+  }
 `;
 
 const Selector = styled.select`
-    margin-left: 5px;
-    width: 92px;
-    height: 30px;
-    border: 1px solid ${theme.border};
-    border-radius: 15px;
-    text-align: center;
-    > option {
-        &:hover {
-            background-color: ${theme.grayBgColor};
-        }
+  margin-left: 5px;
+  width: 92px;
+  height: 30px;
+  border: 1px solid ${theme.border};
+  border-radius: 15px;
+  text-align: center;
+  > option {
+    &:hover {
+      background-color: ${theme.grayBgColor};
     }
+  }
 `;
 
 const RightWrapper = styled.div`
-    margin-right: 1%;
-    > button {
-        width: 124px;
-        height: 40px;
-        border: 1px solid ${theme.white};
-        font-size: 14px;
-        font-weight: 600;
-        line-height: 16.94px;
-        padding-right: 5px;
-        background-color: ${theme.grayBgColor};
-        cursor: pointer;
-        
-        > svg {
-            margin-right: 3px;
-            line-height: 16.94px;
-            margin-top: 2px;
-        }
+  margin-right: 1%;
+  > button {
+    width: 124px;
+    height: 40px;
+    border: 1px solid ${theme.white};
+    font-size: 14px;
+    font-weight: 600;
+    line-height: 16.94px;
+    padding-right: 5px;
+    background-color: ${theme.grayBgColor};
+    cursor: pointer;
+
+    > svg {
+      margin-right: 3px;
+      line-height: 16.94px;
+      margin-top: 2px;
     }
+  }
 `;
 
 const ProductListContainer = styled.div`
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
 `;
 
 const ProductItem = styled.div`
-    display: flex;
-    flex-direction: column;
-    padding: 2px;
-    margin: 10px;
-    width: calc(100% / ${(props) => props.productsPerRow} - 20px);
-    box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  padding: 2px;
+  margin: 10px;
+  width: calc(100% / ${(props) => props.productsPerRow} - 20px);
+  box-sizing: border-box;
 `;
 
 const ProductImage = styled.img`
@@ -283,27 +284,27 @@ const ProductImage = styled.img`
 `;
 
 const ProductInfoWrapper = styled.div`
-    display: flex;
+  display: flex;
 `;
 
 const ProductName = styled.div`
-    font-size: 17px;
-    font-weight: 500;
-    line-heigh: 19.36px;
-    margin-bottom: 5px;
+  font-size: 17px;
+  font-weight: 500;
+  line-heigh: 19.36px;
+  margin-bottom: 5px;
 `;
 
 const ProductPrice = styled.div`
-    font-size: 17px;
-    font-weight: 600;
-    line-heigh: 16.94px;
-    margin-bottom: 5px;
+  font-size: 17px;
+  font-weight: 600;
+  line-heigh: 16.94px;
+  margin-bottom: 5px;
 `;
 
 const ProductDescription = styled.div`
-    font-size: 14px;
-    font-weight: 500;
-    line-heigh: 16.94px;
+  font-size: 14px;
+  font-weight: 500;
+  line-heigh: 16.94px;
 `;
 
 const IconWrapper = styled.div`
@@ -313,60 +314,14 @@ const IconWrapper = styled.div`
 `;
 
 const Icon = styled.img`
-    background-color: ${(props) => (props.isActive ? "#EB4646" : "#f4f4f4")};
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+  background-color: ${(props) => (props.isActive ? "#EB4646" : "#f4f4f4")};
 `;
 
 const PaginationContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 20px;
 `;
 
 const PageButton = styled.button`
